rename video.tsx to root.tsx
